Improve syntax helper for get more than 1 rule
