+++ src/helpers/index.js
@@ -2,22 +2,13 @@
 
 export { default as compose } from './compose';
 export { default as composeAsync } from './composeAsync';
-
 export { default as config } from './config';
-
 export { default as factoryActions } from './factory.actions';
 export { default as factoryIntents } from './factory.intents';
-
 export { default as intersect } from './intersect';
-
-export { default as match } from './match';
-
 export { default as relation } from './relation';
-
 export { default as request } from './request';
-
 export { default as resolve } from './resolve';
-
 export { default as store } from './store';
-
+export { default as syntax } from './syntax';
 export { default as timeout } from './timeout';
