+++ src/intents/intent.translate.js
@@ -1,18 +1,17 @@
 'use strict';
 
-import { factoryActions, intersect, match, resolve } from '../helpers'
+import { factoryActions, resolve, syntax } from '../helpers'
 // -- Internal
-const TERMS = [
-  'translate'
+const RULES = [
+  'translate [Preposition]? [Demonym]',
+  'translate * [Preposition]? [Demonym]',
 ];
-const MATCH_RULE = 'translate [Preposition]? [Demonym]';
 
 export default (state, actions) => {
-  const tokens = intersect(TERMS, state.tokens);
-  const matchRule = match(state.sentence, MATCH_RULE);
-  console.log('IntentTranslate'.bold.green, 'tokens:'.bold, tokens, 'matchRule:'.bold, matchRule);
+  const match = syntax(state.sentence, RULES);
+  console.log('IntentTranslate'.bold.green, 'match:'.bold, match);
 
-  if (tokens || matchSentence) {
+  if (match) {
     return factoryActions(state, actions);
   } else {
     return resolve(state);
